separate validator layout to be shareable
