+++ src/components/Validator/CheckRow.tsx
