+++ src/components/Validator/LogsSection.tsx
