+++ src/views/Generator/ValidatorContent.tsx
@@ -1,17 +1,10 @@
-import { css } from '@emotion/react'
-import { Box, Tabs } from '@radix-ui/themes'
-import { Allotment } from 'allotment'
 import { ReactNode, useEffect, useState } from 'react'
 
-import { ReadOnlyEditor } from '@/components/Monaco/ReadOnlyEditor'
+import { ValidatorLayout } from '@/components/Validator/ValidatorLayout'
 import { HttpRequestDetails } from '@/components/WebLogView/HttpRequestDetails'
 import { useProxyDataGroups } from '@/hooks/useProxyDataGroups'
-import { useTrackScriptCopy } from '@/hooks/useTrackScriptCopy'
 import { K6Check, K6Log, ProxyData } from '@/types'
 import { RequestsSection } from '@/views/Recorder/RequestsSection'
-
-import { ChecksSection } from '../Validator/ChecksSection'
-import { LogsSection } from '../Validator/LogsSection'
 
 interface ValidatorContentProps {
   script: string
@@ -22,8 +15,6 @@ interface ValidatorContentProps {
   checks: K6Check[]
 }
 
-type ValidatorTabValue = 'logs' | 'checks' | 'script'
-
 export function ValidatorContent({
   script,
   proxyData,
@@ -33,7 +24,6 @@ export function ValidatorContent({
   noDataElement,
 }: ValidatorContentProps) {
   const [selectedRequest, setSelectedRequest] = useState<ProxyData | null>(null)
-  const [selectedTab, setSelectedTab] = useState<ValidatorTabValue>('script')
   const groups = useProxyDataGroups(proxyData)
 
   useEffect(() => {
@@ -47,96 +37,29 @@ export function ValidatorContent({
     }
   }, [isRunning])
 
-  useEffect(() => {
-    return window.studio.script.onScriptFailed(() => {
-      setSelectedTab('logs')
-    })
-  }, [])
-
-  const handleCopy = useTrackScriptCopy(script)
+  const details = selectedRequest && (
+    <HttpRequestDetails
+      selectedRequest={selectedRequest}
+      onSelectRequest={setSelectedRequest}
+    />
+  )
 
   return (
-    <Allotment defaultSizes={[3, 2]}>
-      <Allotment.Pane minSize={250}>
-        <Allotment vertical defaultSizes={[1, 1]}>
-          <Allotment.Pane>
-            <RequestsSection
-              proxyData={proxyData}
-              autoScroll={isRunning}
-              selectedRequestId={selectedRequest?.id}
-              noDataElement={noDataElement}
-              onSelectRequest={setSelectedRequest}
-              groups={groups}
-            />
-          </Allotment.Pane>
-          <Allotment.Pane minSize={250}>
-            <Box height="100%">
-              <Tabs.Root
-                value={selectedTab}
-                onValueChange={(value) =>
-                  setSelectedTab(value as ValidatorTabValue)
-                }
-                css={css`
-                  height: 100%;
-                  display: flex;
-                  flex-direction: column;
-                `}
-              >
-                <Tabs.List
-                  css={css`
-                    flex-shrink: 0;
-                  `}
-                >
-                  <Tabs.Trigger value="logs">Logs ({logs.length})</Tabs.Trigger>
-                  <Tabs.Trigger value="checks" disabled={checks.length === 0}>
-                    Checks ({checks.length})
-                  </Tabs.Trigger>
-                  <Tabs.Trigger value="script">Script</Tabs.Trigger>
-                </Tabs.List>
-
-                <Tabs.Content
-                  value="logs"
-                  css={css`
-                    flex: 1;
-                    min-height: 0;
-                  `}
-                >
-                  <LogsSection logs={logs} autoScroll={isRunning} />
-                </Tabs.Content>
-                <Tabs.Content
-                  value="script"
-                  css={css`
-                    flex: 1;
-                  `}
-                >
-                  <ReadOnlyEditor
-                    language="javascript"
-                    value={script}
-                    onCopy={handleCopy}
-                  />
-                </Tabs.Content>
-                <Tabs.Content
-                  value="checks"
-                  css={css`
-                    flex: 1;
-                    min-height: 0;
-                  `}
-                >
-                  <ChecksSection checks={checks} isRunning={isRunning} />
-                </Tabs.Content>
-              </Tabs.Root>
-            </Box>
-          </Allotment.Pane>
-        </Allotment>
-      </Allotment.Pane>
-      {selectedRequest !== null && (
-        <Allotment.Pane minSize={300}>
-          <HttpRequestDetails
-            selectedRequest={selectedRequest}
-            onSelectRequest={setSelectedRequest}
-          />
-        </Allotment.Pane>
-      )}
-    </Allotment>
+    <ValidatorLayout
+      isRunning={isRunning}
+      script={script}
+      logs={logs}
+      checks={checks}
+      details={details}
+    >
+      <RequestsSection
+        proxyData={proxyData}
+        autoScroll={isRunning}
+        selectedRequestId={selectedRequest?.id}
+        noDataElement={noDataElement}
+        onSelectRequest={setSelectedRequest}
+        groups={groups}
+      />
+    </ValidatorLayout>
   )
 }
